+++ bin/coverage_doctest.py
@@ -33,7 +33,7 @@ except ImportError:
 from sympy.utilities.misc import filldedent
 
 
-# Load color templates, used from sympy/utilities/runtests.py
+# Load color templates, duplicated from sympy/testing/runtests.py
 color_templates = (
     ("Black", "0;30"),
     ("Red", "0;31"),
