+++ doc/src/modules/testing/index.rst
@@ -1,0 +1,18 @@
+.. _utilities-docs:
+
+=======
+Testing
+=======
+
+.. TODO: add  benchmarking.rst
+
+.. automodule:: sympy.testing
+
+Contents:
+
+.. toctree::
+   :maxdepth: 2
+
+   pytest.rst
+   randtest.rst
+   runtests.rst
