+++ sympy/testing/__init__.py
@@ -1,0 +1,7 @@
+"""This module contains code for running the tests in SymPy.
+"""
+from .runtests import test, doctest
+
+__all__ = [
+    'test', 'doctest',
+]
