+++ sympy/testing/pytest.py
