+++ sympy/testing/quality_unicode.py
