+++ sympy/testing/tmpfiles.py
